+++ src/include/S3/S3ListPartsRequest.h
@@ -33,10 +33,10 @@
  *
  * The default value is 1000.
  */
-@property (nonatomic) int maxParts;
+@property (nonatomic) long maxParts;
 
 /** The part after which to start the listing. */
-@property (nonatomic) int partNumberMarker;
+@property (nonatomic) long partNumberMarker;
 
 /** Initializes the instance with the bucket, key, and uploadId from an S3MultipartUpload instance returned in an S3MultipartUploadResponse. */
 -(id)initWithMultipartUpload:(S3MultipartUpload *)multipartUpload;
